cleanup derive
